integration: move turn_off_schema_fetching test
It belongs to `metadata/configuration.rs`.
