+++ scylla/tests/integration/metadata/configuration.rs
@@ -1,11 +1,13 @@
 //! Tests that the driver enforces the provided custom metadata fetching timeout
 //! iff ScyllaDB is the target node (else ignores the custom timeout).
 
-use crate::utils::{create_new_session_builder, unique_keyspace_name, PerformDDL as _};
-use crate::utils::{setup_tracing, test_with_3_node_cluster};
+use crate::utils::{
+    create_new_session_builder, setup_tracing, test_with_3_node_cluster, unique_keyspace_name,
+    PerformDDL as _,
+};
 use scylla::client::session::Session;
 use scylla::client::session_builder::SessionBuilder;
-use scylla::cluster::metadata::{ColumnType, NativeType, UserDefinedType};
+use scylla::cluster::metadata::{ColumnType, NativeType, Strategy, UserDefinedType};
 use scylla_proxy::{
     Condition, Reaction as _, RequestFrame, RequestOpcode, RequestReaction, RequestRule,
     ShardAwareness,
@@ -244,4 +246,78 @@ async fn test_refresh_metadata_after_schema_agreement() {
             ])
         }))
     );
+}
+
+#[tokio::test]
+async fn test_turning_off_schema_fetching() {
+    setup_tracing();
+    let session = create_new_session_builder()
+        .fetch_schema_metadata(false)
+        .build()
+        .await
+        .unwrap();
+    let ks = unique_keyspace_name();
+
+    session
+        .ddl(format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks))
+        .await
+        .unwrap();
+
+    session
+        .query_unpaged(format!("USE {}", ks), &[])
+        .await
+        .unwrap();
+
+    session
+        .ddl(
+            "CREATE TYPE IF NOT EXISTS type_a (
+                    a map<frozen<list<int>>, text>,
+                    b frozen<map<frozen<list<int>>, frozen<set<text>>>>
+                   )",
+        )
+        .await
+        .unwrap();
+
+    session
+        .ddl("CREATE TYPE IF NOT EXISTS type_b (a int, b text)")
+        .await
+        .unwrap();
+
+    session
+        .ddl("CREATE TYPE IF NOT EXISTS type_c (a map<frozen<set<text>>, frozen<type_b>>)")
+        .await
+        .unwrap();
+
+    session
+        .ddl(
+            "CREATE TABLE IF NOT EXISTS table_a (
+                    a frozen<type_a> PRIMARY KEY,
+                    b type_b,
+                    c frozen<type_c>,
+                    d map<text, frozen<list<int>>>,
+                    e tuple<int, text>
+                  )",
+        )
+        .await
+        .unwrap();
+
+    session.refresh_metadata().await.unwrap();
+    let cluster_state = &session.get_cluster_state();
+    let keyspace = cluster_state.get_keyspace(&ks).unwrap();
+
+    let datacenter_repfactors: HashMap<String, usize> = cluster_state
+        .replica_locator()
+        .datacenter_names()
+        .iter()
+        .map(|dc_name| (dc_name.to_owned(), 1))
+        .collect();
+
+    assert_eq!(
+        keyspace.strategy,
+        Strategy::NetworkTopologyStrategy {
+            datacenter_repfactors
+        }
+    );
+    assert_eq!(keyspace.tables.len(), 0);
+    assert_eq!(keyspace.user_defined_types.len(), 0);
 }
