+++ scylla/tests/integration/session/session.rs
@@ -7,7 +7,6 @@ use scylla::client::caching_session::CachingSession;
 use scylla::client::execution_profile::ExecutionProfile;
 use scylla::client::session::Session;
 use scylla::client::session_builder::SessionBuilder;
-use scylla::cluster::metadata::Strategy::NetworkTopologyStrategy;
 use scylla::errors::OperationType;
 use scylla::errors::{
     BadKeyspaceName, DbError, ExecutionError, RequestAttemptError, UseKeyspaceError,
@@ -1577,80 +1576,6 @@ async fn test_prepared_config() {
 
     assert!(prepared_statement.get_is_idempotent());
     assert_eq!(prepared_statement.get_page_size(), 42);
-}
-
-#[tokio::test]
-async fn test_turning_off_schema_fetching() {
-    setup_tracing();
-    let session = create_new_session_builder()
-        .fetch_schema_metadata(false)
-        .build()
-        .await
-        .unwrap();
-    let ks = unique_keyspace_name();
-
-    session
-        .ddl(format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks))
-        .await
-        .unwrap();
-
-    session
-        .query_unpaged(format!("USE {}", ks), &[])
-        .await
-        .unwrap();
-
-    session
-        .ddl(
-            "CREATE TYPE IF NOT EXISTS type_a (
-                    a map<frozen<list<int>>, text>,
-                    b frozen<map<frozen<list<int>>, frozen<set<text>>>>
-                   )",
-        )
-        .await
-        .unwrap();
-
-    session
-        .ddl("CREATE TYPE IF NOT EXISTS type_b (a int, b text)")
-        .await
-        .unwrap();
-
-    session
-        .ddl("CREATE TYPE IF NOT EXISTS type_c (a map<frozen<set<text>>, frozen<type_b>>)")
-        .await
-        .unwrap();
-
-    session
-        .ddl(
-            "CREATE TABLE IF NOT EXISTS table_a (
-                    a frozen<type_a> PRIMARY KEY,
-                    b type_b,
-                    c frozen<type_c>,
-                    d map<text, frozen<list<int>>>,
-                    e tuple<int, text>
-                  )",
-        )
-        .await
-        .unwrap();
-
-    session.refresh_metadata().await.unwrap();
-    let cluster_state = &session.get_cluster_state();
-    let keyspace = cluster_state.get_keyspace(&ks).unwrap();
-
-    let datacenter_repfactors: HashMap<String, usize> = cluster_state
-        .replica_locator()
-        .datacenter_names()
-        .iter()
-        .map(|dc_name| (dc_name.to_owned(), 1))
-        .collect();
-
-    assert_eq!(
-        keyspace.strategy,
-        NetworkTopologyStrategy {
-            datacenter_repfactors
-        }
-    );
-    assert_eq!(keyspace.tables.len(), 0);
-    assert_eq!(keyspace.user_defined_types.len(), 0);
 }
 
 #[tokio::test]
